Fixed a few issues including not stopping ad
Removed javascript:pingTracker due to it causing a browser error and not closing the ad.
Removed the code to overwrite the MOTD file due to base triggers hooking the player typing "motd" and opening the url set in the motd file, with no way to stop the ad after. However, the code always forces the pinion ad to load from the config file on connect anyway, so it still works fine and allows the server owner to set what MOTD the users sees after typing "motd".